Updated heuristics to be based on node degree

+++ src/main/java/intsy/group7/pathfinder_sim/model/Edge.java
@@ -3,9 +3,9 @@ package intsy.group7.pathfinder_sim.model;
 public class Edge {
     private Node origin;
     private Node destination;
-    private double weight;
+    private int weight;
 
-public Edge(Node origin, Node destination, double weight){
+public Edge(Node origin, Node destination, int weight){
     this.origin = origin;
     this.destination = destination;
     this.weight = weight;
@@ -19,7 +19,7 @@ public Node getDestination() {
     return this.destination;
 }
 
-public double getWeight() {
+public int getWeight() {
     return this.weight;
 }
 }
